RH-30 : Card 문서화 테스트 분리
